Week9/day5_nexus_ai/requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
autogen-core>=0.4.0
openai>=1.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
